requirements.txt
pandas==2.2.2
pyarrow==17.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
pytest==8.3.2
